use int insteaf uint64 for epoch

+++ api/vdisk/vdisk_api_RollbackVdisk.go
@@ -54,7 +54,7 @@ func (api VdisksAPI) RollbackVdisk(w http.ResponseWriter, r *http.Request) {
 
 	// Create rollback blueprint
 	bp := struct {
-		Timestamp uint64 `yaml:"timestamp" json:"timestamp"`
+		Timestamp int `yaml:"timestamp" json:"timestamp"`
 	}{
 		Timestamp: reqBody.Epoch,
 	}
